Fixed some warnings SD
git-svn-id: svn+ssh://savannah01.psi.ch/repos/gems/GEMSELEKTOR/trunk@944 7d9feeff-4400-0410-95ed-9cb8b92a518d

+++ gems2/vizor/dlg/GraphDialog.cpp
@@ -363,7 +363,7 @@ void GraphDialog::ShowPlots()
            nScale = gr_data.getColorLine( ii );
 
            if( pnt.IsEmpty() || nLn == 0 ||
-               abs(nLn) > gr_data.lines.GetCount() ||
+               (uint)abs(nLn) > gr_data.lines.GetCount() ||
                nScale < 0 )
              continue;   // undefined point or point type
